requirements.txt
ipykernel
langchain
python-dotenv
langchain-community
pypdf
bs4
arxiv
PyMuPDF
wikipedia

langchain-text-splitters

sentence_transformers
langchain_huggingface

faiss-cpu
langchain_chroma

langchain_groq
langchain_core


fastapi
uvicorn

langserve